# huff0: Use static decompression buffer up to 30% faster